pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "job_search"
version = "0.1.0"
description = "Job search tools"
authors = [{ name = "Alexander Wu" }]

readme = "README.md"
classifiers = ["Programming Language :: Python :: 3"]
dependencies = [
    "aw",
    "docx2pdf>=0.1.8",
    "fastapi[standard]>=0.121.0",
    "google-genai>=1.43.0",
    "ipykernel>=7.0.0",
    "ipywidgets>=8.1.7",
    "lxml>=6.0.2",
    "markdown-it-py[plugins]>=4.0.0",
    "markdownify>=1.2.0",
    "mkdocs",
    "mkdocs-material>=9.6.21",
    "mkdocstrings[python]>=0.30.1",
    "pandas>=2.3.3",
    "pip",
    "plotly>=6.3.1",
    "polars>=1.34.0",
    "pyarrow>=21.0.0",
    "pytest",
    "python-docx>=1.2.0",
    "python-dotenv",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "ruff",
    "scipy>=1.16.2",
    "seaborn>=0.13.2",
    "selenium>=4.36.0",
    "tqdm>=4.67.1",
    "webdriver-manager>=4.0.2",
]
requires-python = ">=3.12"


[tool.ruff]
line-length = 99
src = ["job_search"]
include = ["pyproject.toml", "job_search/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"] # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["job_search"]
force-sort-within-sections = true

[tool.uv.sources]
aw = { git = "https://github.com/alexanderwu/aw.git" }
